Django/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.9.0
filters==1.3.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
openapi-codec==1.3.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
unicorn==2.0.1.post1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
whitenoise==6.3.0